show all licences / authorities in table, not just those with applications last week
