+++ spec/javascripts/fakeapi.js
@@ -5,7 +5,8 @@ define([
   'fakeapi/licensing/applications-detail-lastweek',
   'fakeapi/licensing/applications-top5authorities-weekly',
   'fakeapi/licensing/applications-top5licences-weekly',
-  'fakeapi/licensing/applications-total-weekly'
+  'fakeapi/licensing/applications-total-weekly',
+  'fakeapi/licensing/all-entities'
 ],
 function (_) {
   // use XMLHttpRequest for cross domain mock calls on IE
