+++ spec/javascripts/fakeapi/licensing/applications-detail-lastweek.js
@@ -28,11 +28,10 @@ define([
     },
     
     getAuthorities: function (query) {
-      var values = _.map(this.authorities, function (name) {
+      var values = _.map(this.authorities, function (slug) {
         return {
-          authorityUrlSlug: name.replace(' ', '-').toLowerCase(),
+          authorityUrlSlug: slug,
           _count: Math.floor(this.rnd(0, 1e4)),
-          authorityName: [name]
         }
       }, this);
       
@@ -40,11 +39,10 @@ define([
     },
     
     getLicences: function (query) {
-      var values = _.map(this.licences, function (name) {
+      var values = _.map(this.licences, function (slug) {
         return {
-          licenceUrlSlug: name.replace(' ', '-').toLowerCase(),
-          _count: Math.floor(this.rnd(0, 1e4)),
-          licenceName: [name]
+          licenceUrlSlug: slug,
+          _count: Math.floor(this.rnd(0, 1e4))
         }
       }, this);
       
@@ -52,553 +50,245 @@ define([
     },
     
     authorities: [
-      "Aberdeen City Council",
-      "Aberdeenshire Council",
-      "Amber Valley Borough Council",
-      "Angus Council",
-      "Ards Borough Council",
-      "Argyll and Bute Council",
-      "Armagh City and District Council",
-      "Arun District Council",
-      "Ashfield District Council",
-      "Aylesbury Vale District Council",
-      "Babergh District Council",
-      "Ballymena Borough Council",
-      "Ballymoney Borough Council",
-      "Barnsley Metropolitan Borough Council",
-      "Barrow in Furness Borough Council",
-      "Basildon District Council",
-      "Basingstoke and Deane Borough Council",
-      "Bedford Borough Council",
-      "Belfast Health and Social Care trust",
-      "Birmingham City Council",
-      "Blaby District Council",
-      "Blackburn with Darwen Borough Council",
-      "Blackpool Borough Council",
-      "Blaenau Gwent County Borough Council",
-      "Bolsover District Council",
-      "Borough Council of King's Lynn and West Norfolk",
-      "Boston Borough Council",
-      "Bracknell Forest Borough Council",
-      "Bradford Metropolitan District Council",
-      "Braintree District Council",
-      "Breckland Council",
-      "Brentwood Borough Council",
-      "Bridgend County Borough Council",
-      "Brighton and Hove City Council",
-      "Bristol City Council",
-      "Bromsgrove District Council",
-      "Broxbourne Borough Council",
-      "Broxtowe Borough Council",
-      "Burnley Borough Council",
-      "Bury Metropolitan Borough Council",
-      "Caerphilly County Borough Council",
-      "Calderdale Metropolitan Borough Council",
-      "Cambridge City Council",
-      "Cannock Chase District Council",
-      "Canterbury City Council",
-      "Cardiff County Council",
-      "Carlisle City Council",
-      "Carrickfergus Borough Council",
-      "Castle Point Borough Council",
-      "Central Bedfordshire Council",
-      "Ceredigion County Council",
-      "Charnwood Borough Council",
-      "Chelmsford City Council",
-      "Cheltenham Borough Council",
-      "Cherwell District Council",
-      "Cheshire East",
-      "Cheshire West and Chester",
-      "Chesterfield Borough Council",
-      "Chiltern District Council",
-      "Chorley Borough Council",
-      "Christchurch Borough Council",
-      "City and County of Swansea",
-      "City of Edinburgh Council",
-      "City of London",
-      "City of York Council",
-      "Clackmannanshire Council",
-      "Colchester Borough Council",
-      "Coleraine Borough Council",
-      "Corby Borough Council",
-      "Cornwall Council",
-      "Cotswold District Council",
-      "Coventry City Council",
-      "Crawley Borough Council",
-      "Dacorum Borough Council",
-      "Dartford Borough Council",
-      "Daventry District Council",
-      "Denbighshire County Council",
-      "Department For Education",
-      "Department of Agriculture and Rural Development",
-      "Derbyshire County Council",
-      "Derbyshire Dales District Council",
-      "Doncaster Metropolitan Borough Council",
-      "Driver and Vehicle Licensing Agency",
-      "Dungannon and South Tyrone Borough Council",
-      "Durham County Council",
-      "East Ayrshire Council",
-      "East Devon District Council",
-      "East Dorset District Council",
-      "East Dunbartonshire Council",
-      "East Hampshire District Council",
-      "East Hertfordshire District Council",
-      "East Northamptonshire District Council",
-      "East Renfrewshire Council",
-      "East Staffordshire Borough Council",
-      "Eastbourne Borough Council",
-      "Eastleigh Borough Council",
-      "Epping Forest District Council",
-      "Epsom and Ewell Borough Council",
-      "Erewash Borough Council",
-      "Essex County Council",
-      "Exeter City Council",
-      "Falkirk Council",
-      "Fareham Borough Council",
-      "Fenland District Council",
-      "Flintshire County Council",
-      "Food Standards Agency",
-      "Forest Heath District Council",
-      "Forest of Dean District Council",
-      "Gateshead Metropolitan Borough Council",
-      "GDS Test Authority",
-      "GDS Test Authority 2",
-      "GDS Test Authority 3",
-      "Gedling Borough Council",
-      "Gloucester City Council",
-      "Gloucestershire County Council",
-      "Gosport Borough Council",
-      "Gravesham Borough Council",
-      "Great Yarmouth Borough Council",
-      "Guildford Borough Council",
-      "Halton Borough Council",
-      "Harborough District Council",
-      "Harrogate Borough Council",
-      "Hart District Council",
-      "Hartlepool Borough Council",
-      "Havant Borough Council",
-      "Herefordshire Council",
-      "Hertsmere Borough Council",
-      "High Peak Borough Council",
-      "Huntingdonshire District Council",
-      "Inverclyde Council",
-      "Ipswich Borough Council",
-      "Isle of Anglesey County Council",
-      "Kent County Council",
-      "Kettering Borough Council",
-      "Kingston upon Hull City Council",
-      "Kirklees Council",
-      "Lancaster City Council",
-      "Larne Borough Council",
-      "Leeds City Council",
-      "Leicester City Council",
-      "Lewes District Council",
-      "Lichfield District Council",
-      "London Borough of Barking and Dagenham",
-      "London Borough of Barnet",
-      "London Borough of Bexley",
-      "London Borough of Bromley",
-      "London Borough of Croydon",
-      "London Borough of Ealing",
-      "London Borough of Enfield",
-      "London Borough of Hammersmith & Fulham",
-      "London Borough of Haringey",
-      "London Borough of Harrow",
-      "London Borough of Havering",
-      "London Borough of Hillingdon",
-      "London Borough of Hounslow",
-      "London Borough of Islington",
-      "London Borough of Lambeth",
-      "London Borough of Lewisham",
-      "London Borough of Merton",
-      "London Borough of Newham",
-      "London Borough of Redbridge",
-      "London Borough of Richmond upon Thames",
-      "London Borough of Southwark",
-      "London Borough of Sutton",
-      "London Borough of Tower Hamlets",
-      "London Borough of Waltham Forest",
-      "London Borough of Wandsworth",
-      "London Borough of Westminster",
-      "Luton Borough Council",
-      "Maidstone Borough Council",
-      "Malvern Hills District Council",
-      "Medway Council",
-      "Merthyr Tydfil County Borough Council",
-      "Mid Suffolk District Council",
-      "Middlesbrough Council",
-      "Milton Keynes Council",
-      "Mole Valley District Council",
-      "Monmouthshire County Council",
-      "Moray Council",
-      "Neath Port Talbot County Council",
-      "New Forest District Council",
-      "Newark and Sherwood District Council",
-      "Newcastle under Lyme Borough Council",
-      "Newport City Council",
-      "Norfolk County Council",
-      "North Devon District Council",
-      "North East Derbyshire District Council",
-      "North East Lincolnshire Council",
-      "North Hertfordshire District Council",
-      "North Kesteven District Council",
-      "North Norfolk District Council",
-      "North Somerset Council",
-      "North Tyneside Council",
-      "North Warwickshire Borough Council",
-      "North West Leicestershire District Council",
-      "Northampton Borough Council",
-      "Northern Health and Social Care Trust",
-      "Northern Ireland Environment Agency",
-      "Northumberland County Council",
-      "Norwich City Council",
-      "Nottingham City Council",
-      "Nuneaton and Bedworth Borough Council",
-      "Oadby and Wigston Borough Council",
-      "Office for Standards in Education, Children's Services and Skills",
-      "Oldham Metropolitan Borough Council",
-      "Oxford City Council",
-      "Oxfordshire County Council",
-      "Perth and Kinross Council",
-      "Peterborough City Council",
-      "Plymouth City Council",
-      "Poole Borough Council",
-      "Portsmouth City Council",
-      "Powys County Council",
-      "Preston City Council",
-      "Purbeck District Council",
-      "Reading Borough Council",
-      "Redcar & Cleveland Borough Council",
-      "Redditch Borough Council",
-      "Reigate and Banstead Borough Council",
-      "Rhondda Cynon Taff County Borough Council",
-      "Ribble Valley Borough Council",
-      "Rochford District Council",
-      "Rother District Council",
-      "Royal Borough of Greenwich",
-      "Royal Borough of Kensington and Chelsea",
-      "Rugby Borough Council",
-      "Runnymede Borough Council",
-      "Rushcliffe Borough Council",
-      "Rushmoor Borough Council",
-      "Ryedale District Council",
-      "Sandwell Metropolitan Borough Council",
-      "Scarborough Borough Council",
-      "Sedgemoor District Council",
-      "Sefton Metropolitan Borough Council",
-      "Selby District Council",
-      "Sevenoaks District Council",
-      "Sheffield City Council",
-      "Shropshire Council",
-      "Solihull Metropolitan Borough Council",
-      "Somerset County Council",
-      "South Ayrshire Council",
-      "South Bucks District Council",
-      "South Cambridgeshire District Council",
-      "South Derbyshire District Council",
-      "South Gloucestershire Council",
-      "South Hams District Council",
-      "South Kesteven District Council",
-      "South Northamptonshire Council",
-      "South Oxfordshire District Council",
-      "South Ribble Borough Council",
-      "South Staffordshire District Council",
-      "South Tyneside Council",
-      "Southampton City Council",
-      "Spelthorne Borough Council",
-      "St Albans District Council",
-      "St Edmundsbury Borough Council",
-      "Stafford Borough Council",
-      "Staffordshire Moorlands District Council",
-      "Stevenage Borough Council",
-      "Stockton-on-Tees Borough Council",
-      "Stoke-on-Trent City Council",
-      "Stratford on Avon District Council",
-      "Stroud District Council",
-      "Sunderland City Council",
-      "Surrey County Council",
-      "Surrey Heath Borough Council",
-      "Swale Borough Council",
-      "Tamworth Borough Council",
-      "Tandridge District Council",
-      "Taunton Deane Borough Council",
-      "Teignbridge District Council",
-      "Telford & Wrekin Council",
-      "Tendring District Council",
-      "Test Valley Borough Council",
-      "Thanet District Council",
-      "Three Rivers District Council",
-      "Thurrock Borough Council",
-      "Tonbridge and Malling Borough Council",
-      "Torbay Council",
-      "Trafford Council",
-      "Tunbridge Wells Borough Council",
-      "Vale of Glamorgan County Council",
-      "Vale of White Horse District Council",
-      "Wakefield City Metropolitan District Council",
-      "Walsall Council",
-      "Warwick District Council",
-      "Watford Borough Council",
-      "Waverley Borough Council",
-      "Wealden District Council",
-      "Wellingborough Borough Council",
-      "Welwyn Hatfield District Council",
-      "West Devon Borough Council",
-      "West Dorset District Council",
-      "West Lancashire District Council",
-      "West Lindsey District Council",
-      "West Lothian Council",
-      "West Oxfordshire District Council",
-      "West Somerset Council",
-      "Weymouth and Portland Borough Council",
-      "Wigan Council",
-      "Wiltshire County Council",
-      "Winchester City Council",
-      "Wirral Metropolitan Borough Council",
-      "Woking Borough Council",
-      "Wolverhampton City Council",
-      "Worcester City Council",
-      "Worcestershire County Council",
-      "Worthing Borough Council",
-      "Wrexham County Borough Council",
-      "Wychavon District Council",
-      "Wycombe District Council",
-      "Wyre Borough Council",
-      "Wyre Forest District Council"
+      "aberdeen-city-council",
+      "aberdeenshire-council",
+      "amber-valley-borough-council",
+      "angus-council",
+      "ards-borough-council",
+      "argyll-and-bute-council",
+      "armagh-city-and-district-council",
+      "arun-district-council",
+      "ashfield-district-council",
+      "aylesbury-vale-district-council",
+      "babergh-district-council",
+      "ballymena-borough-council",
+      "ballymoney-borough-council",
+      "barnsley-metropolitan-borough-council",
+      "barrow-in-furness-borough-council",
+      "basildon-district-council",
+      "basingstoke-and-deane-borough-council",
+      "bedford-borough-council",
+      "belfast-health-and-social-care-trust",
+      "birmingham-city-council",
+      "blaby-district-council",
+      "blackburn-with-darwen-borough-council",
+      "blackpool-borough-council",
+      "blaenau-gwent-county-borough-council",
+      "bolsover-district-council",
+      "borough-council-of-king's-lynn-and-west-norfolk",
+      "boston-borough-council",
+      "bracknell-forest-borough-council",
+      "bradford-metropolitan-district-council",
+      "braintree-district-council",
+      "breckland-council",
+      "brentwood-borough-council",
+      "bridgend-county-borough-council",
+      "brighton-and-hove-city-council",
+      "bristol-city-council",
+      "bromsgrove-district-council",
+      "broxbourne-borough-council",
+      "broxtowe-borough-council",
+      "burnley-borough-council",
+      "bury-metropolitan-borough-council",
+      "caerphilly-county-borough-council",
+      "calderdale-metropolitan-borough-council",
+      "cambridge-city-council",
+      "cannock-chase-district-council",
+      "canterbury-city-council",
+      "cardiff-county-council",
+      "carlisle-city-council",
+      "carrickfergus-borough-council",
+      "castle-point-borough-council",
+      "central-bedfordshire-council",
+      "ceredigion-county-council",
+      "charnwood-borough-council",
+      "chelmsford-city-council",
+      "cheltenham-borough-council",
+      "cherwell-district-council",
+      "cheshire-east",
+      "cheshire-west-and-chester",
+      "chesterfield-borough-council",
+      "chiltern-district-council",
+      "chorley-borough-council",
+      "christchurch-borough-council",
+      "city-and-county-of-swansea",
+      "city-of-edinburgh-council",
+      "city-of-london",
+      "city-of-york-council",
+      "clackmannanshire-council",
+      "colchester-borough-council",
+      "coleraine-borough-council",
+      "corby-borough-council",
+      "cornwall-council",
+      "cotswold-district-council",
+      "coventry-city-council",
+      "crawley-borough-council",
+      "dacorum-borough-council",
+      "dartford-borough-council",
+      "daventry-district-council",
+      "denbighshire-county-council",
+      "department-for-education",
+      "department-of-agriculture-and-rural-development",
+      "derbyshire-county-council",
+      "derbyshire-dales-district-council",
+      "doncaster-metropolitan-borough-council",
+      "driver-and-vehicle-licensing-agency",
+      "dungannon-and-south-tyrone-borough-council",
+      "durham-county-council",
+      "east-ayrshire-council",
+      "east-devon-district-council",
+      "east-dorset-district-council",
+      "east-dunbartonshire-council",
+      "east-hampshire-district-council",
+      "east-hertfordshire-district-council",
+      "east-northamptonshire-district-council",
+      "east-renfrewshire-council",
+      "east-staffordshire-borough-council",
+      "eastbourne-borough-council",
+      "eastleigh-borough-council",
+      "epping-forest-district-council",
+      "epsom-and-ewell-borough-council",
+      "erewash-borough-council",
+      "essex-county-council",
+      "exeter-city-council",
+      "falkirk-council",
+      "fareham-borough-council",
+      "fenland-district-council",
+      "flintshire-county-council",
+      "food-standards-agency",
+      "forest-heath-district-council",
+      "forest-of-dean-district-council",
+      "gateshead-metropolitan-borough-council",
+      "maidstone-borough-council",
+      "malvern-hills-district-council",
+      "medway-council",
+      "merthyr-tydfil-county-borough-council"
     ],
     
     licences: [
-      "Ability to place tables and chairs in the road",
-      "Actuaries",
-      "Adult placement",
-      "Adult placement service",
-      "Adventure Activities Licence",
-      "Advocate",
-      "Air Conditioning System Energy Assessor",
-      "Air travel organiser's licence (ATOL)",
-      "Animal boarding establishment licence",
-      "Animal boarding establishments",
-      "Animal by-products approval",
-      "Animal Movement",
-      "Animal movement (cattle) notification",
-      "Animal movement notification",
-      "Animal transport authorisation (long/short)",
-      "Animal transport, long journey vehicle approval",
-      "Application for Business Boat Licence",
-      "Application for Registration of an Independent School",
-      "Application for the Childcare Register: childcare provider on domestic or non-domestic premises",
-      "Application for the Childcare Register: childminder",
-      "Application for the Childcare Register: home childcarer",
-      "Application for the Early Years Register: childcare provider on domestic or non-domestic premises",
-      "Application for the Early Years Register: childminder",
-      "Application to register as a children's social care provider: adoption support agency",
-      "Application to register as a children's social care provider: children's home",
-      "Application to register as a children's social care provider: independent fostering service",
-      "Application to register as a children's social care provider: residential family centre",
-      "Application to register as a children's social care provider: voluntary adoption agency",
-      "Application to use street or pavement space for displays",
-      "Approval of a centre to offer SQA qualifications",
-      "Approval of course leading to external qualification",
-      "Approval of Home Child Carers",
-      "Approval of independent school inspector",
-      "Approval of independent school inspectorates",
-      "Approval of meat plants with veterinary attendance",
-      "Approval of premises for civil marriage",
-      "Approval of premises for civil marriage and civil partnership",
-      "Approved Driving Instructor",
-      "Architect",
-      "Assembly Centre Approval",
-      "Auction premises registration",
-      "Audit registration",
-      "Authorisation for a firm to undertake UK audit work",
-      "Authorisation to undertake audit work",
-      "Barrister",
-      "Boat hire licence",
-      "Builder, competent person accreditor",
-      "Buildings: Indvidual and corporate approved inspectors",
-      "Busking licence",
-      "Camp site licence",
-      "Car boot sale authorisation",
-      "Caravan and camping site licence",
-      "Care at home services",
-      "Certificate enabling a premises to offer tourist accommodation",
-      "Certificate of fitness for properties built before 1945",
-      "Charitable collection permit",
-      "Chemical supply and storage (pesticide spraying etc)",
-      "Childminder",
-      "Claims management",
-      "Club licensing",
-      "Commercial (Non-Domestic) Energy Assessor",
-      "Company registration",
-      "Consents for leaflet distribution",
-      "Conveyancers",
-      "Cooling tower notification",
-      "Day care facilities for under 12s",
-      "Day nurseries",
-      "Declaration of compliance with NIEA greenfield soils guidance 1",
-      "Depositing material at sea licence (FEPA)",
-      "Direct appointment as an approved certifier of design or construction",
-      "Discharge of trade effluent permit",
-      "Dispensation to allow the display of fireworks late at night",
-      "Disposal of dredged material at sea",
-      "Diving a designated wreck",
-      "Domestic Energy Assessor - Existing Buildings",
-      "Domiciliary care agencies",
-      "Electricity Distribution Licence",
-      "Electricity Generation Licence",
-      "Electricity Supply Licence",
-      "Electricity Transmission Licence",
-      "Electricty Interconnector",
-      "Emissions trading scheme",
-      "Energy assessor",
-      "Entertainment",
-      "Environmental permitting",
-      "Environmental permitting (national level)",
-      "Estate agency",
-      "Executry practitioners",
-      "Fellow of the Institute of Legal Executives",
-      "Food premises approval",
-      "Food premises registration",
-      "Gas appliances or fittings work",
-      "Gas Interconnector",
-      "Gas shipper",
-      "Gas supplier",
-      "Gas transporter",
-      "Generation, transmission, distribution and supply of electricity",
-      "Guard dog kennels",
-      "Hairdresser licence",
-      "Hairdresser registration",
-      "Hazardous Waste Disposal",
-      "Hold, keep or release any of the specified species of non-native fish",
-      "House in multiple occupancy",
-      "House in multiple occupation licence",
-      "House to house collection",
-      "House to house collection licence",
-      "Hypnotism permit",
-      "Independent school (children's home)",
-      "Independent school standards",
-      "Indoor sports entertainment licence",
-      "Insolvency practitioner",
-      "Insolvency practitioner authorisation",
-      "Insolvency practitioner licence",
-      "Integrated Pollution Prevention Control (IPPC) permit",
-      "Itinerant metal dealer licence",
-      "Knife dealer licence",
-      "Land drainage consents",
-      "Laser or intense pulsed light treatment",
-      "Late hours catering licence",
-      "Law costs draftsmen",
-      "Licence to buy and sell salmon, trout, pike, eel or pollan",
-      "Licence to dispose of waste in the sea",
-      "Licence to kill or take certain birds and animals",
-      "Licence to kill or take certain birds and animals outside 12 mile limit",
-      "Licensed Conveyancers",
-      "Licenses and consents for structures over, along and under the highway",
-      "Listed veterinary nurse",
-      "Local air pollution control authorisation (LAPPC), Part C",
-      "Marine construction works, land reclamation,beach replenishment licence (FEPA)",
-      "Marine works consent",
-      "Market Operator",
-      "Market operators licence",
-      "Market stall licence",
-      "Massage and special treatment premises licensing",
-      "Massage and special treatment, therapist registration",
-      "Metal dealer licence",
-      "Miscellaneous Payments",
-      "Money laundering registration by professional bodies",
-      "Money laundering registration of estate agents",
-      "MOT testing authorisation",
-      "Motor salvage operator registration",
-      "National registration scheme for private landlords",
-      "New Marina Application",
-      "Nurse agency",
-      "Nursing agency",
-      "Occasional licence",
-      "Occupation of the road in connection with building work",
-      "On-Construction Domestic Energy Assessor (OCDEA)",
-      "Operational Ratings Assessor (Production of Display Energy Certificates)",
-      "Packaging waste registration",
-      "Performing animals registration",
-      "Permission to place tables and chairs in the road",
-      "Personal licence to sell alcohol",
-      "Pet shop licence",
-      "Petroleum storage licence",
-      "Pleasure boats",
-      "Pollution prevention and control(PPC)- Part A",
-      "Pollution prevention and control(PPC)- Part B",
-      "Postal operators",
-      "PPC - Dry cleaning",
-      "PPC - Petrol Vapour Recovery",
-      "PPC - Solvent emissions",
-      "Premises licence",
-      "Processing personal data",
-      "Protected species licensing",
-      "Protected wreck site licences",
-      "Public entertainment licence",
-      "Quarantine kennel",
-      "Raised structures",
-      "Register as Number Plate Suppliers",
-      "Registration as a children's home",
-      "Registration carrier / broker of controlled waste",
-      "Registration of Farriers",
-      "Registration of Independent school",
-      "Reporting of pollution inventory data",
-      "Reporting of pollution inventory data (part of EPR)",
-      "Residential care homes",
-      "Residential family centre",
-      "Riding establishment",
-      "Riding establishment licence",
-      "Road/street works operative",
-      "Road/street works supervisor",
-      "Safety certificates for regulated stands at sports grounds",
-      "Safety certificates for sports grounds",
-      "Scheme for certificate of design or construction",
-      "Scheme provider for certificate of design or construction",
-      "School approved for special needs",
-      "Scrap metal dealer registration",
-      "Second hand dealers",
-      "Second hand dealers registration",
-      "Sex establishment / sex cinema",
-      "Sex shop",
-      "Sex shop and cinema licence",
-      "Skin piercing and tattooing",
-      "Skip licence",
-      "Skip operator licence",
-      "Slaughterman",
-      "Social landlords",
-      "Solicitor",
-      "Special school approval",
-      "Special Waste Disposal",
-      "Street collection",
-      "Street collection licence",
-      "Street trader's licence",
-      "Street trading licence",
-      "Streetworks and running pipes/cables under road",
-      "Sunday trading",
-      "Support services providing day care to adults",
-      "Tattooists, piercing and electrolysis licence",
-      "Temporary event notice",
-      "Test Licence",
-      "Theatre licence",
-      "Transmissible spongiform encephalopathies testing",
-      "Transportation & supply of water",
-      "Travelling circus animal welfare",
-      "Venison dealer licence",
-      "Veterinary surgeon",
-      "Waste broker",
-      "Waste carrier",
-      "Waste carrier or broker",
-      "Waste electrical and electronic equipment accreditation  scheme",
-      "Waste management",
-      "Waste management licence",
-      "Waste packaging accreditation",
-      "Water discharge consent",
-      "Water discharge consents",
-      "Water environment controlled activities (CAR) consents",
-      "Weighbridge operator certificate",
-      "Wildlife licence",
-      "Window cleaner",
-      "Zoo licence"
+      "ability-to-place-tables-and-chairs-in-the-road",
+      "actuaries",
+      "adult-placement",
+      "adult-placement-service",
+      "adventure-activities-licence",
+      "advocate",
+      "air-conditioning-system-energy-assessor",
+      "air-travel-organiser's-licence-(atol)",
+      "animal-boarding-establishment-licence",
+      "animal-boarding-establishments",
+      "animal-by-products-approval",
+      "animal-movement",
+      "animal-movement-(cattle)-notification",
+      "animal-movement-notification",
+      "animal-transport-authorisation-(long/short)",
+      "animal-transport,-long-journey-vehicle-approval",
+      "application-for-business-boat-licence",
+      "application-for-registration-of-an-independent-school",
+      "application-for-the-childcare-register:-childcare-provider-on-domestic-or-non-domestic-premises",
+      "application-for-the-childcare-register:-childminder",
+      "application-for-the-childcare-register:-home-childcarer",
+      "application-for-the-early-years-register:-childcare-provider-on-domestic-or-non-domestic-premises",
+      "application-for-the-early-years-register:-childminder",
+      "application-to-register-as-a-children's-social-care-provider:-adoption-support-agency",
+      "application-to-register-as-a-children's-social-care-provider:-children's-home",
+      "application-to-register-as-a-children's-social-care-provider:-independent-fostering-service",
+      "application-to-register-as-a-children's-social-care-provider:-residential-family-centre",
+      "application-to-register-as-a-children's-social-care-provider:-voluntary-adoption-agency",
+      "application-to-use-street-or-pavement-space-for-displays",
+      "approval-of-a-centre-to-offer-sqa-qualifications",
+      "approval-of-course-leading-to-external-qualification",
+      "approval-of-home-child-carers",
+      "approval-of-independent-school-inspector",
+      "approval-of-independent-school-inspectorates",
+      "approval-of-meat-plants-with-veterinary-attendance",
+      "approval-of-premises-for-civil-marriage",
+      "approval-of-premises-for-civil-marriage-and-civil-partnership",
+      "approved-driving-instructor",
+      "architect",
+      "assembly-centre-approval",
+      "auction-premises-registration",
+      "audit-registration",
+      "authorisation-for-a-firm-to-undertake-uk-audit-work",
+      "authorisation-to-undertake-audit-work",
+      "barrister",
+      "boat-hire-licence",
+      "builder,-competent-person-accreditor",
+      "buildings:-indvidual-and-corporate-approved-inspectors",
+      "busking-licence",
+      "camp-site-licence",
+      "car-boot-sale-authorisation",
+      "caravan-and-camping-site-licence",
+      "care-at-home-services",
+      "certificate-enabling-a-premises-to-offer-tourist-accommodation",
+      "certificate-of-fitness-for-properties-built-before-1945",
+      "charitable-collection-permit",
+      "chemical-supply-and-storage-(pesticide-spraying-etc)",
+      "childminder",
+      "claims-management",
+      "club-licensing",
+      "commercial-(non-domestic)-energy-assessor",
+      "company-registration",
+      "consents-for-leaflet-distribution",
+      "conveyancers",
+      "cooling-tower-notification",
+      "day-care-facilities-for-under-12s",
+      "day-nurseries",
+      "declaration-of-compliance-with-niea-greenfield-soils-guidance-1",
+      "depositing-material-at-sea-licence-(fepa)",
+      "direct-appointment-as-an-approved-certifier-of-design-or-construction",
+      "discharge-of-trade-effluent-permit",
+      "dispensation-to-allow-the-display-of-fireworks-late-at-night",
+      "disposal-of-dredged-material-at-sea",
+      "diving-a-designated-wreck",
+      "domestic-energy-assessor---existing-buildings",
+      "domiciliary-care-agencies",
+      "electricity-distribution-licence",
+      "electricity-generation-licence",
+      "electricity-supply-licence",
+      "electricity-transmission-licence",
+      "electricty-interconnector",
+      "emissions-trading-scheme",
+      "energy-assessor",
+      "entertainment",
+      "environmental-permitting",
+      "environmental-permitting-(national-level)",
+      "estate-agency",
+      "executry-practitioners",
+      "fellow-of-the-institute-of-legal-executives",
+      "food-premises-approval",
+      "food-premises-registration",
+      "gas-appliances-or-fittings-work",
+      "gas-interconnector",
+      "gas-shipper",
+      "gas-supplier",
+      "gas-transporter",
+      "generation,-transmission,-distribution-and-supply-of-electricity",
+      "guard-dog-kennels",
+      "hairdresser-licence",
+      "hairdresser-registration",
+      "hazardous-waste-disposal",
+      "hold,-keep-or-release-any-of-the-specified-species-of-non-native-fish",
+      "house-in-multiple-occupancy",
+      "house-in-multiple-occupation-licence",
+      "house-to-house-collection",
+      "house-to-house-collection-licence",
+      "hypnotism-permit",
+      "independent-school-(children's-home)",
+      "independent-school-standards",
+      "indoor-sports-entertainment-licence",
+      "insolvency-practitioner",
+      "insolvency-practitioner-authorisation",
+      "insolvency-practitioner-licence",
+      "integrated-pollution-prevention-control-(ippc)-permit",
+      "itinerant-metal-dealer-licence",
+      "knife-dealer-licence",
+      "land-drainage-consents",
+      "laser-or-intense-pulsed-light-treatment",
+      "late-hours-catering-licence",
+      "law-costs-draftsmen",
+      "licence-to-buy-and-sell-salmon,-trout,-pike,-eel-or-pollan",
+      "licence-to-dispose-of-waste-in-the-sea",
+      "licence-to-kill-or-take-certain-birds-and-animals"
     ]
     
   });
